+++ src/core/plugins/core-commands/client/src/consoleCommands.ts
