+++ src/core/plugins/core-commands/server/config/commandsConfig.ts
