+++ src/core/plugins/core-weather/server/src/commands/weatherCommands.ts
